+++ lib/mix/test/mix/cli_test.exs
@@ -240,6 +240,7 @@ defmodule Mix.CLITest do
     System.delete_env("MIX_EXS")
   end
 
+  @tag :cover
   @tag tmp_dir: "new_with_tests"
   test "new with tests and cover", %{tmp_dir: tmp_dir} do
     File.cd!(tmp_dir, fn ->
